RPackage Golem (#6)
Package the CFDE GitHub Helper app as an RPackage, using the golem
framework. This version may be installed and run locally as any RPackage
would.

To test this PR:
1. checkout this branch
2. run `devtools::install('utilities/CFDE.HelpR')`

After install, run `CFDE.HelpR::run_app()` which will launch the
application. If run through RStudio, it should open in a new browser
window, but if not you'll need to head to http://localhost:8100 for
authentication with GitHub.